+++ src/Microsoft.DotNet.XHarness.Apple/Orchestration/TestOrchestrator.cs
@@ -4,6 +4,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.DotNet.XHarness.Common;
@@ -12,18 +13,24 @@ using Microsoft.DotNet.XHarness.Common.Logging;
 using Microsoft.DotNet.XHarness.iOS.Shared;
 using Microsoft.DotNet.XHarness.iOS.Shared.Execution;
 using Microsoft.DotNet.XHarness.iOS.Shared.Hardware;
+using Microsoft.DotNet.XHarness.iOS.Shared.Listeners;
 using Microsoft.DotNet.XHarness.iOS.Shared.Logging;
+using Microsoft.DotNet.XHarness.iOS.Shared.Utilities;
+using Microsoft.DotNet.XHarness.iOS.Shared.XmlResults;
 
 namespace Microsoft.DotNet.XHarness.Apple
 {
     /// <summary>
-    /// This orchestrator implements the `test` command flow.
-    /// In this flow we install, run and uninstall the application.
-    /// We need to connect to the running application over TCP and receive the test results.
-    /// We also need to watch timeouts better and parse the results more comprehensively.
+    /// Common ancestor for `test` and `just-test` orchestrators.
     /// </summary>
-    public class TestOrchestrator : BaseTestOrchestrator
+    public abstract class TestOrchestrator : BaseOrchestrator
     {
+        private readonly IMlaunchProcessManager _processManager;
+        private readonly ILogger _logger;
+        private readonly ILogs _logs;
+        private readonly IFileBackedLog _mainLog;
+        private readonly IErrorKnowledgeBase _errorKnowledgeBase;
+
         public TestOrchestrator(
             IMlaunchProcessManager processManager,
             IAppBundleInformationParser appBundleInformationParser,
@@ -32,8 +39,13 @@ namespace Microsoft.DotNet.XHarness.Apple
             ILogs logs,
             IFileBackedLog mainLog,
             IErrorKnowledgeBase errorKnowledgeBase)
-            : base(processManager, appBundleInformationParser, deviceFinder, consoleLogger, logs, mainLog, errorKnowledgeBase)
+            : base(processManager, appBundleInformationParser, deviceFinder, consoleLogger, mainLog, errorKnowledgeBase)
         {
+            _processManager = processManager ?? throw new ArgumentNullException(nameof(processManager));
+            _logger = consoleLogger ?? throw new ArgumentNullException(nameof(consoleLogger));
+            _logs = logs ?? throw new ArgumentNullException(nameof(logs));
+            _mainLog = mainLog ?? throw new ArgumentNullException(nameof(mainLog));
+            _errorKnowledgeBase = errorKnowledgeBase ?? throw new ArgumentNullException(nameof(errorKnowledgeBase));
         }
 
         public Task<ExitCode> OrchestrateTest(
@@ -90,6 +102,146 @@ namespace Microsoft.DotNet.XHarness.Apple
                 executeMacCatalystApp,
                 executeApp,
                 cancellationToken);
+        }
+
+        private async Task<ExitCode> ExecuteApp(
+            AppBundleInformation appBundleInfo,
+            TestTargetOs target,
+            IDevice device,
+            IDevice? companionDevice,
+            TimeSpan timeout,
+            TimeSpan launchTimeout,
+            CommunicationChannel communicationChannel,
+            XmlResultJargon xmlResultJargon,
+            IEnumerable<string> singleMethodFilters,
+            IEnumerable<string> classMethodFilters,
+            IReadOnlyCollection<(string, string)> environmentalVariables,
+            IEnumerable<string> passthroughArguments,
+            CancellationToken cancellationToken)
+        {
+            AppTester appTester = GetAppTester(communicationChannel, target.Platform.IsSimulator());
+
+            (TestExecutingResult testResult, string resultMessage) = await appTester.TestApp(
+                appBundleInfo,
+                target,
+                device,
+                companionDevice,
+                timeout,
+                launchTimeout,
+                passthroughArguments,
+                environmentalVariables,
+                xmlResultJargon,
+                skippedMethods: singleMethodFilters?.ToArray(),
+                skippedTestClasses: classMethodFilters?.ToArray(),
+                cancellationToken: cancellationToken);
+
+            return ParseResult(testResult, resultMessage);
+        }
+
+        private async Task<ExitCode> ExecuteMacCatalystApp(
+            AppBundleInformation appBundleInfo,
+            TimeSpan timeout,
+            TimeSpan launchTimeout,
+            CommunicationChannel communicationChannel,
+            XmlResultJargon xmlResultJargon,
+            IEnumerable<string> singleMethodFilters,
+            IEnumerable<string> classMethodFilters,
+            IReadOnlyCollection<(string, string)> environmentalVariables,
+            IEnumerable<string> passthroughArguments,
+            CancellationToken cancellationToken)
+        {
+            AppTester appTester = GetAppTester(communicationChannel, TestTarget.MacCatalyst.IsSimulator());
+
+            (TestExecutingResult testResult, string resultMessage) = await appTester.TestMacCatalystApp(
+                appBundleInfo,
+                timeout,
+                launchTimeout,
+                passthroughArguments,
+                environmentalVariables,
+                xmlResultJargon,
+                skippedMethods: singleMethodFilters?.ToArray(),
+                skippedTestClasses: classMethodFilters?.ToArray(),
+                cancellationToken: cancellationToken);
+
+            return ParseResult(testResult, resultMessage);
+        }
+
+        private AppTester GetAppTester(CommunicationChannel communicationChannel, bool isSimulator)
+        {
+            var tunnelBore = (communicationChannel == CommunicationChannel.UsbTunnel && !isSimulator)
+                ? new TunnelBore(_processManager)
+                : null;
+
+            // Only add the extra callback if we do know that the feature was indeed enabled
+            Action<string>? logCallback = IsLldbEnabled() ? (l) => NotifyUserLldbCommand(_logger, l) : null;
+
+            return new AppTester(
+                _processManager,
+                new SimpleListenerFactory(tunnelBore),
+                new CrashSnapshotReporterFactory(_processManager),
+                new CaptureLogFactory(),
+                new DeviceLogCapturerFactory(_processManager),
+                new TestReporterFactory(_processManager),
+                new XmlResultParser(),
+                _mainLog,
+                _logs,
+                new Helpers(),
+                logCallback: logCallback);
+        }
+
+        private ExitCode ParseResult(TestExecutingResult testResult, string resultMessage)
+        {
+            string newLine = Environment.NewLine;
+            const string checkLogsMessage = "Check logs for more information";
+
+            void LogProblem(string message)
+            {
+                if (_errorKnowledgeBase.IsKnownTestIssue(_mainLog, out var issue))
+                {
+                    _logger.LogError(message + newLine + issue.Value.HumanMessage);
+                }
+                else
+                {
+                    if (resultMessage != null)
+                    {
+                        _logger.LogError(message + newLine + resultMessage + newLine + newLine + checkLogsMessage);
+                    }
+                    else
+                    {
+                        _logger.LogError(message + newLine + checkLogsMessage);
+                    }
+                }
+            }
+
+            switch (testResult)
+            {
+                case TestExecutingResult.Succeeded:
+                    _logger.LogInformation($"Application finished the test run successfully");
+                    _logger.LogInformation(resultMessage);
+                    return ExitCode.SUCCESS;
+
+                case TestExecutingResult.Failed:
+                    _logger.LogInformation($"Application finished the test run successfully with some failed tests");
+                    _logger.LogInformation(resultMessage);
+                    return ExitCode.TESTS_FAILED;
+
+                case TestExecutingResult.LaunchFailure:
+                    LogProblem("Failed to launch the application");
+                    return ExitCode.APP_LAUNCH_FAILURE;
+
+                case TestExecutingResult.Crashed:
+                    LogProblem("Application run crashed");
+                    return ExitCode.APP_CRASH;
+
+                case TestExecutingResult.TimedOut:
+                    _logger.LogWarning($"Application test run timed out");
+                    return ExitCode.TIMED_OUT;
+
+                default:
+                    _logger.LogError($"Application test run ended in an unexpected way: '{testResult}'" +
+                        newLine + (resultMessage != null ? resultMessage + newLine + newLine : null) + checkLogsMessage);
+                    return ExitCode.GENERAL_FAILURE;
+            }
         }
     }
 }
